+++ NEWS.md
@@ -6,6 +6,8 @@
 * the log likelihood calculations performed by `logLik` and `loglikelihood` have
   been revised, expanded to include three possible definitions of the likelihood 
   function, and documented in a new vignette
+* `tune_vlmc` and `tune_covlmc` can be used with the different likelihood 
+  function definitions
 * `cutoff` uses a new `tolerance` parameter to avoid reporting cut off values
   that are almost identical due to numerical imprecision